+++ src/hutch_bunny/cli.py
@@ -4,7 +4,7 @@ from hutch_bunny.core.results_modifiers import (
     get_results_modifiers_from_str,
 )
 from hutch_bunny.core.execute_query import execute_query
-from hutch_bunny.core.rquest_dto.result import RquestResult
+from hutch_bunny.core.rquest_models.result import RquestResult
 from hutch_bunny.core.parser import parser
 from hutch_bunny.core.logger import configure_logger, logger
 from hutch_bunny.core.db import get_db_manager
