+++ src/hutch_bunny/core/solvers/availability_solver.py
@@ -36,13 +36,13 @@ from typing import Tuple
 from sqlalchemy import exists
 
 from hutch_bunny.core.obfuscation import apply_filters
-from hutch_bunny.core.rquest_dto.group import Group
-from hutch_bunny.core.rquest_dto.query import AvailabilityQuery
+from hutch_bunny.core.rquest_models.group import Group
+from hutch_bunny.core.rquest_models.query import AvailabilityQuery
 from sqlalchemy.engine import Engine
 from hutch_bunny.core.logger import logger, INFO
 
 from hutch_bunny.core.settings import Settings
-from hutch_bunny.core.rquest_dto.rule import Rule
+from hutch_bunny.core.rquest_models.rule import Rule
 
 
 class ResultModifier(TypedDict):
