+++ src/hutch_bunny/core/solvers/query_solvers.py
@@ -4,11 +4,11 @@ from hutch_bunny.core.logger import logger
 
 from hutch_bunny.core.solvers.availability_solver import AvailabilitySolver
 from hutch_bunny.core.db_manager import SyncDBManager
-from hutch_bunny.core.rquest_dto.query import AvailabilityQuery, DistributionQuery
-from hutch_bunny.core.rquest_dto.file import File
+from hutch_bunny.core.rquest_models.query import AvailabilityQuery, DistributionQuery
+from hutch_bunny.core.rquest_models.file import File
 from hutch_bunny.core.solvers.availability_solver import ResultModifier
 
-from hutch_bunny.core.rquest_dto.result import RquestResult
+from hutch_bunny.core.rquest_models.result import RquestResult
 from hutch_bunny.core.enums import DistributionQueryType
 from hutch_bunny.core.settings import Settings
 from hutch_bunny.core.constants import DISTRIBUTION_TYPE_FILE_NAMES_MAP
