+++ tests/conftest.py
@@ -2,11 +2,11 @@ import pytest
 import os
 
 from hutch_bunny.core.enums import DistributionQueryType
-from hutch_bunny.core.rquest_dto.cohort import Cohort
-from hutch_bunny.core.rquest_dto.group import Group
-from hutch_bunny.core.rquest_dto.rule import Rule
+from hutch_bunny.core.rquest_models.cohort import Cohort
+from hutch_bunny.core.rquest_models.group import Group
+from hutch_bunny.core.rquest_models.rule import Rule
 from hutch_bunny.core.db_manager import SyncDBManager
-from hutch_bunny.core.rquest_dto.query import AvailabilityQuery, DistributionQuery
+from hutch_bunny.core.rquest_models.query import AvailabilityQuery, DistributionQuery
 from hutch_bunny.core.settings import Settings
 import hutch_bunny.core.db as db
