+++ tests/integration/test_demographics_distribution_query.py
@@ -1,7 +1,7 @@
 import pytest
 from hutch_bunny.core.solvers.query_solvers import solve_distribution
-from hutch_bunny.core.rquest_dto.result import RquestResult
-from hutch_bunny.core.rquest_dto.file import File
+from hutch_bunny.core.rquest_models.result import RquestResult
+from hutch_bunny.core.rquest_models.file import File
 
 
 @pytest.fixture
